Modified the list parsers.
Starting to think this is a bit of a dead end. I think I may have to
parse sequentially into a node, and then add the node to the DOM at
the end. It shouldn't matter much but it just seems way more workable.

This means I'll have to loop through all the markdown rules and see
which one matches first, handle it, and repeat both inside the new
node that results, and outside.

It kind of feels wrong to use the DOM and then just modify innerHTML
but I'm not going to overcomplicate my code even more by changing
that.